requirements.txt
whitenoise==6.5.0
uwsgi==2.0.21
psycopg2==2.9.6
phonenumberslite==8.13.15
openpyxl==3.1.2
pandas==2.0.3
numpy==1.25.1
requests==2.31.0
requests-ntlm==1.2.0
scikit-learn==1.3.0
xlsxwriter==3.1.2

# Django
django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-phonenumber-field==7.1.0
django-grappelli==3.0.6
django-leaflet==0.29.0